Update CheckDiskAcccessPermissions.cs
The Try was for the whole procedure, both check for current path and the individual drives.
This meant that when there was no access to to a folder down the tree from the current path, the drives check was skipped, giving a false sense of security. I add two trys now, one for going down the tree and one around each drive check.